+++ tests/unit-tests/Tests/ReadOnlySpan/Linq/Sum.cs
@@ -1,0 +1,12 @@
+﻿using static SpanExtensions.Tests.UnitTests.TestHelper;
+
+namespace SpanExtensions.Tests.UnitTests
+{
+    public static partial class ReadOnlySpanLinqTests
+    {
+        public sealed class Sum
+        {
+            // todo
+        }
+    }
+}
