--- tests/unit-tests/Tests/ReadOnlySpan/ReadOnlySpanTests.String.cs
@@ -1,9 +1,0 @@
-using static SpanExtensions.Tests.UnitTests.TestHelper;
-
-namespace SpanExtensions.Tests.UnitTests
-{
-    public static class ReadOnlySpanStringTests
-    {
-
-    }
-}
