+++ tests/unit-tests/Tests/ReadOnlySpan/Split/SplitAny.cs
@@ -1,0 +1,119 @@
+﻿using static SpanExtensions.Tests.UnitTests.TestHelper;
+
+namespace SpanExtensions.Tests.UnitTests
+{
+    public static partial class ReadOnlySpanSplitTests
+    {
+        public sealed class SplitAny
+        {
+            [Fact]
+            public void ConsecutiveDelimitersResultInEmptySpan()
+            {
+                AssertEqual(
+                    [['a'], [], ['a']],
+                    "abba".AsSpan().SplitAny(['b', 'c']).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a'], [], ['a']],
+                    "abca".AsSpan().SplitAny(['b', 'c']).ToSystemEnumerable()
+                );
+            }
+
+            [Fact]
+            public void DelimiterAtTheEndResultInEmptySpan()
+            {
+                AssertEqual(
+                    [['a', 'a'], []],
+                    "aab".AsSpan().SplitAny(['b', 'c']).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a'], []],
+                    "aac".AsSpan().SplitAny(['b', 'c']).ToSystemEnumerable()
+                );
+            }
+
+            [Fact]
+            public void DelimiterAtTheEndWithCountEqualDelimiterCountResultsInSpanWithDelimiter()
+            {
+                AssertEqual(
+                    [['a', 'a', 'b']],
+                    "aab".AsSpan().SplitAny(['b', 'c'], 1).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a', 'c']],
+                    "aac".AsSpan().SplitAny(['b', 'c'], 1).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a'], ['a', 'c']],
+                    "aabac".AsSpan().SplitAny(['b', 'c'], 2).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a'], ['a', 'b']],
+                    "aacab".AsSpan().SplitAny(['b', 'c'], 2).ToSystemEnumerable()
+                );
+            }
+
+            [Fact]
+            public void CountEqualDelimiterCountResultsInSpanWithEverythingAfterAndIncludingLastDelimiter()
+            {
+                AssertEqual(
+                    [['a', 'a', 'b', 'a', 'a']],
+                    "aabaa".AsSpan().SplitAny(['b', 'c'], 1).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a', 'c', 'a', 'a']],
+                    "aacaa".AsSpan().SplitAny(['b', 'c'], 1).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a'], ['a', 'a', 'c', 'a', 'a']],
+                    "aabaacaa".AsSpan().SplitAny(['b', 'c'], 2).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a'], ['a', 'a', 'b', 'a', 'a']],
+                    "aacaabaa".AsSpan().SplitAny(['b', 'c'], 2).ToSystemEnumerable()
+                );
+            }
+
+            [Fact]
+            public void CountEqualDelimiterCountResultsInEverythingAfterAndIncludingLastDelimiterBeingCut()
+            {
+                AssertEqual(
+                    [['a', 'a']],
+                    "aabaa".AsSpan().SplitAny(['b', 'c'], 1, CountExceedingBehaviour.CutLastElements).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a']],
+                    "aacaa".AsSpan().SplitAny(['b', 'c'], 1, CountExceedingBehaviour.CutLastElements).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a'], ['a', 'a']],
+                    "aabaacaa".AsSpan().SplitAny(['b', 'c'], 2, CountExceedingBehaviour.CutLastElements).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [['a', 'a'], ['a', 'a']],
+                    "aacaabaa".AsSpan().SplitAny(['b', 'c'], 2, CountExceedingBehaviour.CutLastElements).ToSystemEnumerable()
+                );
+            }
+
+            [Fact]
+            public void CountEqualZeroReturnsNothing()
+            {
+                AssertEqual(
+                    [],
+                    "aabc".AsSpan().SplitAny(['b', 'c'], 0, CountExceedingBehaviour.CutLastElements).ToSystemEnumerable()
+                );
+                AssertEqual(
+                    [],
+                    "aabc".AsSpan().SplitAny(['d', 'e'], 0, CountExceedingBehaviour.CutLastElements).ToSystemEnumerable()
+                );
+            }
+
+            [Fact]
+            public void NegativeCountThrowsArgumentOutOfRangeException()
+            {
+                Assert.Throws<ArgumentOutOfRangeException>(() => "aabc".AsSpan().SplitAny(['b', 'c'], -1));
+                Assert.Throws<ArgumentOutOfRangeException>(() => "aabc".AsSpan().SplitAny(['d', 'e'], -1));
+            }
+        }
+    }
+}
