+++ tests/unit-tests/Tests/ReadOnlySpan/String/Remove.cs
@@ -1,0 +1,12 @@
+﻿using static SpanExtensions.Tests.UnitTests.TestHelper;
+
+namespace SpanExtensions.Tests.UnitTests
+{
+    public static partial class ReadOnlySpanStringTests
+    {
+        public sealed class Remove
+        {
+            // todo
+        }
+    }
+}
