+++ tests/unit-tests/Tests/Span/Linq/Sum.cs
@@ -1,0 +1,12 @@
+﻿using static SpanExtensions.Tests.UnitTests.TestHelper;
+
+namespace SpanExtensions.Tests.UnitTests
+{
+    public static partial class SpanLinqTests
+    {
+        public sealed class Sum
+        {
+            // todo
+        }
+    }
+}
