--- tests/unit-tests/Tests/Span/SpanTests.String.cs
@@ -1,9 +1,0 @@
-using static SpanExtensions.Tests.UnitTests.TestHelper;
-
-namespace SpanExtensions.Tests.UnitTests
-{
-    public static class SpanStringTests
-    {
-
-    }
-}
